fix case when there is no note in database with specified id

+++ db.py
@@ -19,21 +19,26 @@ class BadeModel(Model):
 class Note(BadeModel):
     text = TextField()
 
-    # TODO: проверить на случай если записки в бд нет
     @staticmethod
     def get_text_by_id(id: int) -> str:
         """
         Returns text of the note by note's id
         and delete note after that
+
+        If there is no note in database 'None' will be returned
+
         :param id: Id of the note
         :return: str text of the note
         """
-        note = Note.get_by_id(id)
+        note = Note.get_or_none(Note.id == id)
 
-        text = note.text
-        note.delete_instance()
+        if note is None:
+            return None
+        else:
+            text = note.text
+            note.delete_instance()
 
-        return text
+            return text
 
 
 Note.create_table()
